Extract out custom input as component
